début de séparation entre jeu et affichage
potentiellement encore a changer selon les besoins des groupes

+++ src/Plateau.java
@@ -1,5 +1,5 @@
 public class Plateau{
-    int[] plateau;//tableau de cases
+    short[] plateau;//tableau de cases
     int nLignes; //nombre de lignes
     int nColonnes; //nombre de colonnes
     int nCases; //nombre de cases
@@ -8,13 +8,13 @@ public class Plateau{
         this.nLignes = nLignes;
         this.nColonnes = nColonnes;
         this.nCases = nLignes*nColonnes;
-        this.plateau = new int[nCases];
+        this.plateau = new short[nCases];
         init_plateau();
     }
 
     public void init_plateau(){
         for(int i=0; i<nLignes; i++){
-            plateau[i] = nColonnes;
+            plateau[i] = (short) nColonnes;
         }
         nCases = nLignes * nColonnes;
     }
@@ -31,7 +31,7 @@ public class Plateau{
         for(int k = i; k < nLignes; k++){
             if (plateau[k] > j){
                 nCases -= (plateau[k] - j);
-                plateau[k] = j;
+                plateau[k] = (short) j;
             }  
         }
     }
